test: Move conversation ID tests to their own file
